requirements.txt
pandas
matplotlib
seaborn
plotly
spotipy
python-dotenv
scikit-learn
scipy
